fix(sales): cast float8[] au lieu de int[] — quantity est un Float (décimales possibles)
Tous les types corrigés dans $executeRaw :
- id : text[] (Prisma stocke les UUIDs en text, pas uuid)
- qty : float8[] (quantity est number/Float, peut être 0.5, 1.5...)
- table : public.products (@@map + search_path Supabase pooler)

+++ backend/src/commerce/sales/sales.service.ts
@@ -189,7 +189,7 @@ export class SalesService {
         SET    "stockQty" = p."stockQty" - v.qty
         FROM   unnest(
                  ${productIds}::text[],
-                 ${quantities}::int[]
+                 ${quantities}::float8[]
                ) AS v(id, qty)
         WHERE  p.id = v.id
         AND    p."tenantId" = ${tenantId}
